cleaned up routes
deleted unused routes and moved to new files to read better
